+++ .idea/misc.xml
@@ -13,8 +13,9 @@
         <entry key="..\:/Users/123/AndroidStudioProjects/FilmsSearch/app/src/main/res/drawable/ic_baseline_watch_later.xml" value="0.199" />
         <entry key="..\:/Users/123/AndroidStudioProjects/FilmsSearch/app/src/main/res/drawable/ic_round_menu.xml" value="0.199" />
         <entry key="..\:/Users/123/AndroidStudioProjects/FilmsSearch/app/src/main/res/layout/activity_details.xml" value="0.25" />
-        <entry key="..\:/Users/123/AndroidStudioProjects/FilmsSearch/app/src/main/res/layout/activity_main.xml" value="0.1" />
+        <entry key="..\:/Users/123/AndroidStudioProjects/FilmsSearch/app/src/main/res/layout/activity_main.xml" value="0.15" />
         <entry key="..\:/Users/123/AndroidStudioProjects/FilmsSearch/app/src/main/res/layout/film_item.xml" value="0.1" />
+        <entry key="..\:/Users/123/AndroidStudioProjects/FilmsSearch/app/src/main/res/layout/fragment_home.xml" value="0.153125" />
         <entry key="..\:/Users/123/AndroidStudioProjects/FilmsSearch/app/src/main/res/menu/navigation.xml" value="0.20416666666666666" />
         <entry key="..\:/Users/123/AndroidStudioProjects/FilmsSearch/app/src/main/res/menu/navigation_menu.xml" value="0.20416666666666666" />
         <entry key="..\:/Users/123/AndroidStudioProjects/FilmsSearch/app/src/main/res/menu/settings.xml" value="0.203125" />
