+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:theme="@style/Theme.FilmsSearch"
         tools:targetApi="31">
         <activity
-            android:name=".DetailsActivity"
+            android:name=".DetailsFragment"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
