--- app/src/main/java/com/example/FilmsSearch/DetailsActivity.kt
@@ -1,42 +1,0 @@
-package com.example.FilmsSearch
-
-import android.os.Bundle
-import android.widget.Toast
-import androidx.appcompat.app.AppCompatActivity
-import kotlinx.android.synthetic.main.activity_details.*
-
-class DetailsActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setDetailedContent()
-        navigationInit()
-    }
-
-    private fun setDetailedContent() {
-        setContentView(R.layout.activity_details)
-        val film = intent.extras?.get("film") as Film
-        details_toolbar.title = film.title
-        details_poster.setImageResource(film.poster)
-        details_description.text = film.description
-    }
-    private fun navigationInit(){
-        bottom_navigation.setOnNavigationItemSelectedListener {
-            when (it.itemId) {
-                R.id.favorites -> {
-                    Toast.makeText(this, R.string.menu_favorites_title, Toast.LENGTH_SHORT).show()
-                    true
-                }
-                R.id.watch_later -> {
-                    Toast.makeText(this, R.string.menu_watch_later_title, Toast.LENGTH_SHORT).show()
-                    true
-                }
-                R.id.selections -> {
-                 Toast.makeText(this, R.string.menu_selections_title, Toast.LENGTH_SHORT).show()
-                 true
-                }
-                else -> false
-            }
-        }
-    }
-
-}
