+++ app/src/main/res/values-en-rUS/strings.xml
@@ -1,10 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string name="app_name">FilmsSearch</string>
-    <string name="top_text1">Pictures is from https://www.movieposters.com/</string>
     <string name="page_title">Find it!</string>
     <string name="menu_settings_title">Settings</string>
     <string name="menu_favorites_title">Favorites</string>
     <string name="menu_watch_later_title">Watch later</string>
     <string name="menu_selections_title">Selections</string>
+    <string name="disclaimer">Pictures is from https://www.movieposters.com/</string>
+    <string name="double_tap_text">Double tap for exit</string>
 </resources>
