+++ app/src/main/res/values/strings.xml
@@ -1,9 +1,10 @@
 <resources>
     <string name="app_name">FilmsSearch</string>
-    <string name="top_text1">Картинки взяты с https://www.movieposters.com/</string>
+    <string name="disclaimer">Картинки взяты с https://www.movieposters.com/</string>
     <string name="page_title">Найди!</string>
     <string name="menu_settings_title">Настройки</string>
     <string name="menu_favorites_title">Избранное</string>
     <string name="menu_watch_later_title">Просмотреть позже</string>
     <string name="menu_selections_title">Подборки</string>
+    <string name="double_tap_text">Нажмите дважды для выхода</string>
 </resources>
